Add files via upload
Task 1: Create an array of numbers. Use map() to create a new array with each number squared.

Task 2: Use filter() on an array of numbers to get a new array with only even numbers.

Task 3: Create an array of product prices. Use reduce() to find the total price.

Task 4: For an array of strings, use map() to create a new array that contains the length of each string.

Task 5: Use splice() to remove the third item of an array and replace it with the string“replaced!”.